CTD input compoments are placed in a dedicated file components/core/input/ctd_input.
